Report implementation
Report implementation

+++ src/main/java/br/com/dev/simples/patients/dto/AppointmentDTO.java
@@ -1,5 +1,6 @@
 package br.com.dev.simples.patients.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import br.com.dev.simples.patients.model.enumeration.AppointmentType;
@@ -19,8 +20,8 @@ public class AppointmentDTO {
 	private LocalDateTime localDateTime;
 	private AppointmentType type;
 	private String patientId;
-	private double weight;
-	private double waistCircumference;
-	private double glicemia;
+	private BigDecimal weight;
+	private BigDecimal waistCircumference;
+	private BigDecimal glicemia;
 
 }
